Version 0.0.3.65:
- usb2.c auf neue Transaction-Funktionen umgestellt

git-svn-id: svn://svn.code.sf.net/p/prettyos/code/trunk/Source@1266 31957799-4632-4615-9793-72db6d48bcac

+++ user/other_userprogs/music.c
@@ -21,7 +21,7 @@ void PlayTone(unsigned int Frequency, double DurFak) {
 
 void Play(const char* string) {
     unsigned int Frequency = 0;
-    for (int i = 0; string[i] != 0; ++i) {
+    for (size_t i = 0; string[i] != 0; ++i) {
         switch (string[i]) {
             case 'c':
                 Frequency = 132;
